No More GYM and NUMPY
I'm trying to avoid all possible CPU operations, of course it's not entirely possible as every program requires some CPU to be used, but my goal is to use as much of the GPU for training as possible, avoiding as many latency problems as possible.

Note to Self: I might roll this one back, it's not seeming as efficient as gym is, but I could just be impacient.